# spec/kompot.openapi.json
{
  "components": {
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "tacku",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/forms/bulk-move": {
      "get": {
        "operationId": "bulkMoveForm",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "kompot-forms.schema.json#/$defs/KompotFormResponse"
                }
              }
            },
            "description": "ok"
          },
          "401": {
            "description": "no token, or a token issued for another resource"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-kompot-endpoint-kind": "form"
      }
    },
...
    "/forms/new-board": {
      "get": {
        "operationId": "newBoardForm",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "kompot-forms.schema.json#/$defs/KompotFormResponse"
                }
              }
            },
            "description": "ok"
          },
          "401": {
            "description": "no token, or a token issued for another resource"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-kompot-endpoint-kind": "form"
      }
    },
...
    "/submit/bulk-move": {
      "post": {
        "operationId": "submitBulkMove",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "kompot-core.schema.json#/$defs/KompotAction"
                }
              }
            },
            "description": "ok"
          },
          "400": {
            "description": "no idempotency key"
          },
          "401": {
            "description": "no token, or a token issued for another resource"
          },
          "409": {
            "description": "the key was used for a different request"
          },
          "422": {
            "description": "refused on its merits"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-kompot-endpoint-kind": "submit"
      }
    },
...
    "/submit/new-board": {
      "post": {
        "operationId": "submitNewBoard",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "kompot-core.schema.json#/$defs/KompotAction"
                }
              }
            },
            "description": "ok"
          },
          "400": {
            "description": "no idempotency key"
          },
          "401": {
            "description": "no token, or a token issued for another resource"
          },
          "409": {
            "description": "the key was used for a different request"
          },
          "422": {
            "description": "refused on its merits"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-kompot-endpoint-kind": "submit"
      }
    },
...
    "/submit/seen": {
      "post": {
        "operationId": "submitSeen",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "kompot-core.schema.json#/$defs/KompotAction"
                }
              }
            },
            "description": "ok"
          },
          "400": {
            "description": "no idempotency key"
          },
          "401": {
            "description": "no token, or a token issued for another resource"
          },
          "409": {
            "description": "the key was used for a different request"
          },
          "422": {
            "description": "refused on its merits"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-kompot-endpoint-kind": "submit"
      }
    },
...
  },
  "security": [
    {
      "bearer": []
    }
  ],
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ]
}